move mock indexer
